Remove redux listener
