+++ webapp/src/js/controllers/tasks.js
@@ -1,5 +1,3 @@
-var _ = require('underscore');
-
 (function () {
 
   'use strict';
@@ -8,25 +6,21 @@ var _ = require('underscore');
     function(
       $ngRedux,
       $scope,
-      $state,
       $stateParams,
       $timeout,
-      $translate,
       $window,
       GlobalActions,
       LiveList,
       RulesEngine,
       Selectors,
       TasksActions,
-      Tour,
-      TranslateFrom
+      Tour
     ) {
       'ngInject';
 
       const ctrl = this;
       const mapStateToTarget = function(state) {
         return {
-          currentTab: Selectors.getCurrentTab(state),
           selectedTask: Selectors.getSelectedTask(state)
         };
       };
@@ -36,40 +30,9 @@ var _ = require('underscore');
         return {
           unsetSelected: globalActions.unsetSelected,
           setSelectedTask: tasksActions.setSelectedTask,
-          setShowContent: globalActions.setShowContent,
-          setTitle: globalActions.setTitle,
-          settingSelected: globalActions.settingSelected
         };
       };
       const unsubscribe = $ngRedux.connect(mapStateToTarget, mapDispatchToTarget)(ctrl);
-
-      var setSelectedTask = function(task) {
-        LiveList.tasks.setSelected(task._id);
-        ctrl.setSelectedTask(task);
-        if (_.isString(task.title)) {
-          // new translation key style
-          task.title = $translate.instant(task.title, task);
-        } else {
-          // old message array style
-          task.title = TranslateFrom(task.title, task);
-        }
-        ctrl.setTitle(TranslateFrom(task.title, task));
-        ctrl.setShowContent(true);
-      };
-
-      $scope.setSelected = function(id) {
-        if (!id) {
-          LiveList.tasks.clearSelected();
-          ctrl.unsetSelected();
-          return;
-        }
-        var task = _.findWhere(LiveList.tasks.getList(), { _id: id });
-        if (task) {
-          var refreshing = (ctrl.selectedTask && ctrl.selectedTask._id) === id;
-          ctrl.settingSelected(refreshing);
-          setSelectedTask(task);
-        }
-      };
 
       ctrl.refreshTaskList = function() {
         $window.location.reload();
@@ -90,32 +53,24 @@ var _ = require('underscore');
         });
       });
 
-      LiveList.tasks.notifyChange = function(task) {
+      LiveList.tasks.notifyChange = function() {
         ctrl.hasTasks = LiveList.tasks.count() > 0;
-        if (ctrl.selectedTask && task._id === ctrl.selectedTask._id ||
-            (!ctrl.selectedTask && task._id === $state.params.id)) {
-          setSelectedTask(task);
-        }
       };
       LiveList.tasks.notifyError = function() {
         ctrl.error = true;
         ctrl.unsetSelected();
       };
 
-      $scope.$on('query', function() {
-        if (ctrl.currentTab !== 'tasks') {
-          LiveList.tasks.unsetSelected();
-          delete LiveList.tasks.notifyChange;
-          delete LiveList.tasks.notifyError;
-          return;
-        }
-      });
-
       if ($stateParams.tour) {
         Tour.start($stateParams.tour);
       }
 
-      $scope.$on('$destroy', unsubscribe);
+      $scope.$on('$destroy', () => {
+        LiveList.tasks.unsetSelected();
+        delete LiveList.tasks.notifyChange;
+        delete LiveList.tasks.notifyError;
+        unsubscribe();
+      });
     }
   );
